Latest legacy service

+++ services/legacy/database/ffldb/whitebox_test.go
@@ -26,7 +26,7 @@ import (
 )
 
 // fixedExcessiveBlockSize should not be the default -we want to ensure it will work in all cases
-const fixedExcessiveBlockSize uint32 = 42111000
+const fixedExcessiveBlockSize uint64 = 42111000
 
 func init() {
 	wire.SetLimits(fixedExcessiveBlockSize)
