+++ services/legacy/wire/msgfilteradd.go
@@ -67,8 +67,8 @@ func (msg *MsgFilterAdd) Command() string {
 
 // MaxPayloadLength returns the maximum length the payload can be for the
 // receiver.  This is part of the Message interface implementation.
-func (msg *MsgFilterAdd) MaxPayloadLength(pver uint32) uint32 {
-	return uint32(VarIntSerializeSize(MaxFilterAddDataSize)) +
+func (msg *MsgFilterAdd) MaxPayloadLength(pver uint32) uint64 {
+	return uint64(VarIntSerializeSize(MaxFilterAddDataSize)) +
 		MaxFilterAddDataSize
 }
